requirements.txt
# 量化交易系统依赖
pandas>=1.5.0
numpy>=1.21.0
python-binance>=1.0.19
requests>=2.28.0
aiohttp>=3.8.0
websockets>=10.0
pyyaml>=6.0
plotly>=5.10.0
streamlit>=1.20.0
flask>=3.0.0
scipy>=1.10.0

# 测试
pytest>=7.0.0
pytest-cov>=4.0.0